[no ticket] break up e2e cohorts-create test into two tests
